Add files via upload
Hello Carol,

My examples will include 4 of my most recent projects Im continually improving and working on my skills.
Unit 2 is a venomous snake of the world site. It has 4 pages. The assignment demonstrates my ability to use hotpots. You will notice the first 2 hotpots are in Southern California for the western diamond back rattle snake and Eastern Arizona for the Sonora coral snake. The hotpots link to a separate page with information about two venomous species of snakes from that part of the world. The next link can be found in Africa in the middle of the continent and to the right near where Ethiopia is on the map. There you will find a link to the Gaboon viper.  The final fourth link is visible when your move the mouse cursor over the eastern portion of India. There you will find a link for the Russell’s viper.

Unit 3 demonstrates the use of media queries and responsive design. The sight should adapt to look clean for all screen types.

Tutorial 4 is demonstrates how CSS and HTML can be used in graphic design. Like the venomous snake site from the home page, it utilizes hotspots with clickable areas where clicking on the face of each family member will take you to their respective page on the site with information about the family member whose face you clicked. 

Homework 7 demonstrates the use of webforms by implementing a customer satisfaction survey. The form just checks the information entered by the user and validates whether or not the information entered by the user is valid or invalid. 

Thank you for considering my application and taking the time to learn more about me. I enjoyed learning about 44i and think I would be a good fit for your company. Hope you have a fabulous day and I hope to hear from you soon.